F401 and F403 fixes: Simplify admin packages
